o/requirements.txt
Flask
requests
kafka-python